合并最终版可以正常运行但是没有上电机测试不确定是否能使用

+++ src/AlgorithmLayer/rp_math.c
@@ -160,7 +160,7 @@ float step_limit_filter(float new_value, float last_value, float max_step)
 	float difference = new_value - last_value;
 
     // 如果变化量超过最大步进值，则进行限幅步进处理
-	if (fabs(difference) > max_step)
+	if (fabsf(difference) > max_step)
 	{
 		filtered_value = last_value + sgn(difference) * max_step;
 	}
